requirements.txt
chainlit>=1.0.0
litellm>=1.0.0
pydantic>=2.0.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
